% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cars}
\alias{cars}
\title{Cars dataset}
\format{
A data frame with 50 observations and 2 variables:
\describe{
\item{speed}{Speed of the car (mph)}
\item{dist}{Stopping distance (ft)}
}
}
\source{
R built-in dataset
}
\usage{
cars
}
\description{
A dataset containing the speed of cars and the distances taken to stop.
}
\keyword{datasets}
