% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.R
\name{mymean}
\alias{mymean}
\title{Compute the mean of a numeric vector}
\usage{
mymean(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
The mean value of the vector
}
\description{
Compute the mean of a numeric vector
}
